refactor: Swap build package for gradle
