+++ build-logic/convention-plugins/src/main/kotlin/com/omricat/maplibrarian/gradle/ConfigureAndroidKotlin.kt
@@ -1,4 +1,4 @@
-package com.omricat.maplibrarian.build
+package com.omricat.maplibrarian.gradle
 
 import com.android.build.api.dsl.CommonExtension
 import org.gradle.api.Project
@@ -10,7 +10,10 @@ internal fun Project.configureAndroidKotlin(commonExtension: CommonExtension<*, 
     with(commonExtension) {
         compileSdk = compileSdkFromGradleProperties().get()
 
-        defaultConfig.minSdk = minSdkFromGradleProperties().get()
+        defaultConfig {
+            minSdk = minSdkFromGradleProperties().get()
+            testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        }
 
         compileOptions {
             sourceCompatibility = javaVersion
